fix embedding for RNN
