+++ pobax/models/actor_critic.py
@@ -1,0 +1,60 @@
+from flax import linen as nn
+from jax import numpy as jnp
+from jax._src.nn.initializers import orthogonal, constant
+
+from pobax.models import SimpleNN, ScannedRNN, DiscreteActor, ContinuousActor, Critic
+from .embedding import CNN, BattleshipEmbedding
+
+
+class ActorCritic(nn.Module):
+    env_name: str
+    action_dim: int
+    hidden_size: int = 128
+    double_critic: bool = False
+    memoryless: bool = False
+    is_discrete: bool = True
+    is_image: bool = False
+
+    def setup(self):
+        if self.is_image:
+            self.embedding = CNN(hidden_size=self.hidden_size)
+        elif self.env_name == 'battleship':
+            self.embedding = BattleshipEmbedding(hidden_size=self.hidden_size)
+        elif not self.memoryless:
+            self.embedding = nn.Sequential([
+                nn.Dense(self.hidden_size, kernel_init=orthogonal(jnp.sqrt(2)), bias_init=constant(0.0)),
+                nn.relu
+            ])
+        else:
+            self.embedding = SimpleNN(hidden_size=self.hidden_size)
+
+        if not self.memoryless:
+            self.memory = ScannedRNN(hidden_size=self.hidden_size)
+        if self.is_discrete:
+            self.actor = DiscreteActor(self.action_dim, hidden_size=self.hidden_size)
+        else:
+            self.actor = ContinuousActor(self.action_dim, hidden_size=self.hidden_size)
+
+        if self.double_critic:
+            self.critic = nn.vmap(Critic,
+                                  variable_axes={'params': 0},
+                                  split_rngs={'params': True},
+                                  in_axes=None,
+                                  out_axes=2,
+                                  axis_size=2)(hidden_size=self.hidden_size)
+        else:
+            self.critic = Critic(hidden_size=self.hidden_size)
+
+    def __call__(self, hidden, x):
+        obs_dict, dones = x
+        obs = obs_dict.obs
+        action_mask = obs_dict.action_mask
+        embedding = self.embedding(obs)
+        if not self.memoryless:
+            rnn_in = (embedding, dones)
+            hidden, embedding = self.memory(hidden, rnn_in)
+
+        pi = self.actor(embedding, action_mask=action_mask)
+        v = self.critic(embedding)
+
+        return hidden, pi, jnp.squeeze(v, axis=-1)
