+++ pobax/models/discrete.py
@@ -1,16 +1,13 @@
 import distrax
 import flax.linen as nn
-import jax
 import jax.numpy as jnp
 from jax._src.nn.initializers import orthogonal, constant
 import numpy as np
 
 from . import FullImageCNN
-from .network import ScannedRNN, SmallImageCNN
+from .network import SmallImageCNN
 from .value import Critic
 from pobax.models.transformerXL import Transformer
-from .embedding import CNN, BattleshipEmbedding, SimpleNN
-from .continuous import ContinuousActor
 
 
 class DiscreteActor(nn.Module):
@@ -31,51 +28,6 @@ class DiscreteActor(nn.Module):
         pi = distrax.Categorical(logits=actor_mean)
         return pi
 
-class ActorCritic(nn.Module):
-    env_name: str
-    action_dim: int
-    hidden_size: int = 128
-    double_critic: bool = False
-    memoryless: bool = False
-    is_discrete: bool = True
-    is_image: bool = False
-    
-    def setup(self):
-        if self.is_image:
-            self.embedding = CNN(hidden_size=self.hidden_size)
-        elif self.env_name == 'battleship':
-            self.embedding = BattleshipEmbedding(hidden_size=self.hidden_size)
-        else:
-            self.embedding = SimpleNN(hidden_size=self.hidden_size)
-        if not self.memoryless:
-            self.memory = ScannedRNN(hidden_size=self.hidden_size)
-        if self.is_discrete:
-            self.actor = DiscreteActor(self.action_dim, hidden_size=self.hidden_size)
-        else:
-            self.actor = ContinuousActor(self.action_dim, hidden_size=self.hidden_size)
-        if self.double_critic:
-            self.critic = nn.vmap(Critic,
-                             variable_axes={'params': 0},
-                             split_rngs={'params': True},
-                             in_axes=None,
-                             out_axes=2,
-                             axis_size=2)(hidden_size=self.hidden_size)
-        else:
-            self.critic = Critic(hidden_size=self.hidden_size)
-
-    def __call__(self, hidden, x):
-        obs_dict, dones = x
-        obs = obs_dict.obs
-        action_mask = obs_dict.action_mask
-        embedding = self.embedding(obs)
-        if not self.memoryless:
-            rnn_in = (embedding, dones)
-            hidden, embedding = self.memory(hidden, rnn_in)
-        
-        pi = self.actor(embedding, action_mask=action_mask)
-        v = self.critic(embedding)
-
-        return hidden, pi, jnp.squeeze(v, axis=-1)
 
 class DiscreteActorCriticTransformer(nn.Module):
     action_dim: int
